refactor(generate): lift Immich delivery out of the orchestrator
Delivery is the phase that runs after the artifact is already durable:
its whole job is to hand the file to Immich and record delivered,
pending, or failed without ever invalidating a finished render. It has
its own retry contract with RunTracker and a public entry point the UI
calls outside generate_memory(), which makes it a cohesion seam rather
than a line-count cut.

generate.py 882 -> 764; generate_delivery.py is 153 lines. The only
lines added to generate.py are its import block; the moved code differs
from the original by two lines, both routing DeliveryError construction
through a function-scoped import so the new module does not have to
import generate.py at module scope.

GenerationError and DeliveryError stay behind: they are the public
exception surface a dozen callers import from immich_memories.generate.

Refs #685
